2025.10.13 Disto App  - 측정 저장된 데이터 리스트 삭제 기능 추가  - 측정 데이터 리스트 맨홀 타입 오타 수정  - recyclerView margin 수정

+++ app/src/main/java/com/terra/terradisto/ui/survey_diameter/model/SurveyResult.java
@@ -1,12 +1,12 @@
 package com.terra.terradisto.ui.survey_diameter.model;
 
 public class SurveyResult {
-    public long id;
+    public int id;
     public String manholType;
     public double distance;
     public String pipMaterial;
 
-    public SurveyResult(long id, String manholType, double distance, String pipMaterial) {
+    public SurveyResult(int id, String manholType, double distance, String pipMaterial) {
         this.id = id;
         this.manholType = manholType;
         this.distance = distance;
